memmove logic error correction

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: todina-r <todina-r@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/22 09:26:08 by todina-r          #+#    #+#             */
-/*   Updated: 2026/02/04 11:43:18 by todina-r         ###   ########.fr       */
+/*   Updated: 2026/02/05 05:36:17 by todina-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	size_t	index;
+	int	index;
 
 	if (dest < src)
 	{
